Add UAVCAN motor config client in separate class
It turns out that trying to accomodate UAVCAN nodes and ROS nodes
in the same class object is a bad idea (who would have guessed?).
It's more sane to separate the UAVCAN node from the ROS node
and only pass information from one to the other in a controlled manner.